Add Hybrid Mode to Query Adapter
The big change. This adds a check to see if we are running in Hybrid
mode. If we are we look to see which SQL connection strings are
specified. If one is specified we run that portion of XMLDB with SQL
adatpers. Otherwise we run it with the file adapters.