Add the chaarcteradapter fix check
Adds a check to see if we are using the character adapter fix. Its a
hacky way of doing it, but when creating a character in test mode the
server expects account_character_ref to be in test mode as well. This
creates a link to the SQL DB for that one table instead of the file
adapter.